requirements.txt
# 图片处理库
pillow>=10.0.0

# MCP 服务器框架
fastmcp>=0.1.0

# MCP 核心库 (fastmcp 的依赖，但明确列出以确保兼容性)
mcp>=1.0.0
